Apply suggestions from code review

+++ python/cudf/cudf/core/column/column.py
@@ -500,7 +500,7 @@ class ColumnBase(Column, Serializable, NotIterable):
         value_normalized = (
             cudf.Scalar(value, dtype=self.dtype)
             if is_scalar(value)
-            else as_column(value).astype(self.dtype)
+            else as_column(value, dtype=self.dtype)
         )
 
         out: Optional[ColumnBase]  # If None, no need to perform mimic inplace.
@@ -560,8 +560,8 @@ class ColumnBase(Column, Serializable, NotIterable):
             skip_reducing_key = False
             if isinstance(value, ColumnBase):
                 if len(self) == len(value):
-                    # Both value and key is aligned to self. Thus, the values
-                    # corresponding to the `F` masks in key should be ignored.
+                    # Both value and key are aligned to self. Thus, the values
+                    # corresponding to the false values in key should be ignored.
                     value = value.apply_boolean_mask(key)
                     # After applying boolean mask, the length of value equals
                     # the number of elements to scatter, we can skip computing
